requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
pydantic==1.10.5
sqlalchemy==1.4.39
psycopg2-binary==2.9.3
redis==4.5.1
transformers==4.30.0
torch==2.0.0
openai==1.2.0
pytest==7.2.2
pytest-mock==3.10.0
httpx==0.24.1
alembic==1.9.1
python-dotenv==1.0.0
